+++ packages/optimizely-sdk/lib/core/condition_tree_evaluator/index.ts
@@ -14,13 +14,12 @@
  * limitations under the License.                                           *
  ***************************************************************************/
 
+import { ConditionTree } from '../project_config/entities';
 const AND_CONDITION = 'and';
 const OR_CONDITION = 'or';
 const NOT_CONDITION = 'not';
 
 const DEFAULT_OPERATOR_TYPES = [AND_CONDITION, OR_CONDITION, NOT_CONDITION];
-
-export type ConditionTree<Leaf> = Leaf | unknown[];
 
 type LeafEvaluator<Leaf> = (leaf: Leaf) => boolean | null;
